requirements.txt
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.6.0
matplotlib>=3.5.0
numpy>=1.21.0
albumentations>=1.3.0
Pillow>=9.0.0
tqdm>=4.64.0
scikit-learn>=1.1.0
tensorboard>=2.10.0
segmentation-models-pytorch>=0.3.0